requirements.txt
dash==1.20.0
dash-core-components==1.16.0
dash-html-components==1.1.3
dash-leaflet==0.1.13
dash-renderer==1.9.1
dash-extensions==0.0.56
dash-table==4.11.3
geobuf==1.1.1
geopandas==0.9.0
jsbeautifier==1.14.0